+++ jardin/model.py
@@ -4,7 +4,8 @@ import re, inspect
 import json
 
 import jardin.config as config
-from jardin.database import DatabaseAdapter, Datasources
+from jardin.database import DatabaseAdapter
+from jardin.database.datasources import Datasources
 from jardin.tools import soft_del, classorinstancemethod, stack_marker
 from jardin.query import query
