+++ jardin/query.py
@@ -2,7 +2,8 @@ import argparse
 import os
 import inspect
 
-from jardin.database import DatabaseAdapter, Datasources
+from jardin.database import DatabaseAdapter
+from jardin.database.datasources import Datasources
 from jardin.tools import stack_marker
 
 def query(sql=None, filename=None, extract=None, db=None, **kwargs):
